requirements.txt
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
flask==3.1.2
idna==3.10
importlib-metadata==8.7.0
itsdangerous==2.2.0
jinja2==3.1.6
lxml==6.0.2
markupsafe==3.0.2
pillow==11.3.0
pymediainfo==7.0.1
python-pptx==1.0.2
requests==2.32.5
typing-extensions==4.15.0
urllib3==2.5.0
werkzeug==3.1.3
xlsxwriter==3.2.9
zipp==3.23.0
